// src/interfaces/orderbook.ts
export type orderStatus =
  | "NEW"
  | "CANCELLED"
  | "REJECTED"
  | "COMPLETED"
  | "FILLED"
  | "PARTIAL_FILLED"
  | "INCOMPLETE";

export interface OrderlyOrder {
  symbol: string;
  client_order_id?: string;
  order_type: "LIMIT" | "MARKET" | "IOC" | "FOK" | "POST_ONLY" | "ASK" | "BID";
  order_price?: string | number;
  order_quantity?: string | number;
  order_amount?: number | string;
  side: "BUY" | "SELL";
  broker_id?: string;
  visible_quantity?: number;
  reduce_only?: boolean;
}

export interface MyOrder {
  order_id: number;
  user_id: number;
  price: number;
  type: "LIMIT" | "MARKET" | "IOC" | "FOK" | "POST_ONLY" | "ASK" | "BID";
  quantity: number;
  amount?: any;
  executed: number;
  visible: number;
  symbol: string;
  side: "BUY" | "SELL";
  status: orderStatus;
  total_fee: number;
  fee_asset: string;
  client_order_id?: any;
  average_executed_price: number;
  broker_id: string;
  broker_name: string;
  created_time: number;
  updated_time: number;
  reduce_only?: boolean;
}
// ...
export interface Trade {
  symbol: string;
  side: "BUY" | "SELL";
  executed_price: number;
  executed_quantity: number;
  executed_timestamp: number;
}

export interface TokenMetadata {
  id: string;
  name: string;
  symbol: string;
  decimals: number;
  icon: string;
}

export interface Ticker {
  symbol: string;
  open: number;
  close: number;
  high: number;
  low: number;
  volume: number;
  amount: number;
  count: number;
  symbol_from_token_meta?: TokenMetadata;
  symbol_to_token_meta?: TokenMetadata;
}

export interface TokenInfo {
  token: string;
  token_account_id: string;
  decimals: number;
  minimum_increment: number;
}

export interface Holding {
  token: string;
  holding: number;
  frozen: number;
  pending_short: number;
  updated_time: number;
}

export interface ClientInfo {
  account_id: string;
  email: string;
  account_mode: string;
  max_leverage: number;
  tier: string;
  taker_fee_rate: number;
  maker_fee_rate: number;
  futures_taker_fee_rate: number;
  futures_maker_fee_rate: number;
  imr_factor: any;
  maintenance_cancel_orders: boolean;
}
// ...
export interface PositionsType {
  current_margin_ratio_with_orders: number;
  free_collateral: number;
  initial_margin_ratio: number;
  initial_margin_ratio_with_orders: number;
  maintenance_margin_ratio: number;
  maintenance_margin_ratio_with_orders: number;
  margin_ratio: number;
  open_margin_ratio: number;
  rows: PositionsRow[];
  total_collateral_value: number;
  total_pnl_24_h: number;
  timestamp?: number;
}

interface PositionsRow {
  IMR_withdraw_orders: number;
  MMR_with_orders: number;
  average_open_price: number;
  cost_position: number;
  est_liq_price: number;
  fee_24_h: number;
  imr: number;
  last_sum_unitary_funding: number;
  mark_price: number;
  mmr: number;
  pending_long_qty: number;
  pending_short_qty: number;
  pnl_24_h: number;
  position_qty: number;
  settle_price: number;
  symbol: string;
  timestamp: number;
  unsettled_pnl: number;
  display_est_liq_price?: number;
}
// ...
export interface LiquidationType {
  liquidation_id: number;
  timestamp: number;
  transfer_amount_to_insurance_fund: number;
  positions_by_perp: {
    abs_liquidation_fee: number;
    cost_position_transfer: number;
    liquidator_fee: number;
    position_qty: number;
    symbol: string;
    transfer_price: number;
    insurance_fund_fee: number;
  }[];
}